Add function to `entrypoint.sh` to print commands
The functions in the `entrypoint.sh` script have commands written twice,
once to print the command and once to run the command. This is redundant
and potentially error-prone (if only one of the commands were updated).

This commit will add a helper function for printing the commands, and
will replace the duplicate commands with a single helper function call.
Importantly, `printf` output will not be included in the function output
(it is redirected to stderr with `>&2`), so environment variables can
be set from the output of the function, as `DVHE_PROFILE` is currently.
The helper function is mainly useful for matching previous formatting.
Note that it would be simpler to just use `set -x` to trace commands.

https://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#set