Allow configuration of Dolby Vision FEL conversion
The `entrypoint.sh` script used in the `dovi_tool` Docker image converts
Dolby Vision Profile 7 (`dvhe.07.06`) to Profile 8 (`dvhe.08.06`).
In addition to the HDR10 Base Layer (BL), Profile 7 includes an
"Enhancement Layer" (EL) and "Reference Picture Unit" (RPU) information.

The Profile 7 EL comes in two variants:

1. "Minimal Enhancement Layer" (MEL). On MEL sources, the EL is empty
   (included only for compatibility purposes) and the RPU has all of the
   Dolby Vision data.
2. "Full Enhancement Layer" (FEL). On FEL sources, the EL has additional
   color information, as well as luma and chroma mappings, increasing
   color bit depth to 12 bit.

Profile 8 only includes BL+RPU and does not include the EL. Converting a
MEL from Profile 7 to Profile 8 is effectively a lossless conversion
(because the RPU contains all Dolby Vision data), but converting a FEL
from Profile 7 to Profile 8 is a lossy conversion because the additional
color data must be discarded.

It is therefore helpful to know if a Profile 7 source is MEL or FEL.
As FEL conversion is lossy, some users may wish to stop here instead
of converting the FEL and losing picture information.

This commit will add support for the environment variable `CONVERT_FEL`.
If `CONVERT_FEL=0` and a FEL is detected, the `entrypoint.sh` script
will exit prior to converting the source file. The default is `1` to
preserve the previous behavior.

cdf1c28
e59d3b9