Remove unnecessary `RUN`s from Dockerfile
There were two `RUN` commands in the Dockerfile related to installing
`dovi_tool`, one to uncompress the downloaded archive, and one to delete
the downloaded archive. This commit will consolidate these two `RUN`s
into a single `RUN`.

There was also a `RUN` command for creating the `WORKDIR`. This was
unnecessary because `WORKDIR` will automatically create the directory
if it does not exist. This commit will remove the `RUN` command for
creating the `WORKDIR`.
https://docs.docker.com/reference/dockerfile/#workdir

There was another unnecessary `RUN` to make the `entrypoint.sh` script
executable (`chmod +x`). The script is already executable as stored in
the Git repo, but it can sometimes be a good idea to ensure permissions
haven't changed after `COPY`. Dockerfiles now support `COPY --chmod` so
that permissions can be set without a separate `RUN`. This commit will
remove the `RUN` command for making `entrypoint.sh` executable. The
`COPY --chmod` syntax was added in Dockerfile 1.2, so a syntax directive
will be added to ensure the Dockerfile is parsed properly.
https://docs.docker.com/reference/dockerfile/#copy