Fix `dovi_tool` arch
The GitHub Actions workflow builds Docker container images for both
`linux/amd64` (`x86_64`) and `linux/arm64` (`aarch64`). However, the
Dockerfile is currently configured to always install `dovi_tool` built
for `linux/amd64` (`x86_64`). This means the `dovi_tool` installed in
the `linux/arm64` (`aarch64`) images is built for the incorrect arch.

This commit will correct the `dovi_tool` arch so that the correct build
of `dovi_tool` is installed on `linux/arm64` (`aarch64`). Docker offers
automatic platform ARGs in the global scope, including `TARGETARCH`, so
that the current arch can be determined within the Dockerfile. There is
additional programming needed to map between the Docker arch and the
arch specified in the `dovi_tool` download URL. Docker (via containerd)
normalizes arch names, so `x86_64` is always normalized to `amd64` and
`aarch64` is always normalized to `arm64`. The `dovi_tool` download URLs
use the un-normalized arch names. To avoid adding a `RUN` instruction
for setting the arch in the download URL, heredoc/here-doc syntax will
be used. The `ADD` will be replaced by a `wget` command in the heredoc.

https://docs.docker.com/reference/cli/docker/buildx/build/#platform
https://docs.docker.com/reference/dockerfile/#automatic-platform-args-in-the-global-scope
https://github.com/containerd/containerd/blob/v1.4.3/platforms/platforms.go#L63
https://github.com/moby/buildkit/blob/HEAD/frontend/dockerfile/docs/reference.md