Create L1 plots before summarizing RPU
The `summarize_rpu()` function runs before the `create_plot()` function.
The `summarize_rpu()` function supports the `STOP_IF_FEL` environment
variable and will cause the entrypoint script to cleanup and exit if
`STOP_IF_FEL=1` and a FEL is detected. This means L1 plots are not
generated.

This commit will reorder the function invocations so `create_plot()`
will still run and generate plots even if a FEL is detected.

c315e8f
1e69f93