Check for Dolby Vision FEL before conversion
The `entrypoint.sh` script used in the `dovi_tool` Docker image converts
Dolby Vision Profile 7 (`dvhe.07.06`) to Profile 8 (`dvhe.08.06`), and
supports the environment variable `STOP_IF_FEL` for handling Profile 7
FEL (Full Enhancement Layer). FEL conversion is lossy, so some users may
wish to stop here instead of converting the FEL and losing picture info.

As mentioned in 1e69f93, the script always converts the `.hevc` file to
Profile 8 first prior to extracting the RPU layer and retrieving info on
the Enhancement Layer variant (MEL or FEL). While this does work, it is
slower than it needs to be for FEL files when `STOP_IF_FEL=1` is set.
The script will waste time converting to Profile 8 when it is just going
to stop and discard the converted Profile 8 files.

This commit will update the script to check for Profile 7 FEL before
converting to Profile 8. This will be faster when `STOP_IF_FEL=1` is set
because the script will exit before converting to Profile 8. This update
requires changes to how the RPU-related functions are written because
the functions currently assume that both the Profile 7 RPU and Profile 8
RPU are present. Now that the Profile 7 RPU will be analyzed before the
Profile 8 RPU, the functions need to accept a single RPU. This means the
`cleanup "$1"` lines within those functions would actually be using the
RPU file as `$1` instead of the MKV file passed as the first argument to
the `entrypoint.sh` script. To avoid this side effect, `cleanup` calls
will be removed from the RPU-related functions and moved to the function
invocations within `demux_file()` instead.

c315e8f
1e69f93