Exit with error if Dolby Vision profile not found
The `entrypoint.sh` script includes a `get_dvhe_profile()` function that
checks for the supported Dolby Vision profile prior to running profile
conversions. If the supported Profile 7 is not found, the script exits
with `0`. This doesn't make sense because lacking a supported Profile 7
is an error in this context.

This commit will update `get_dvhe_profile()` to exit with `1` instead
of `0` if the supported Profile 7 is not found.