Fix README info on summarizing RPUs
When the `summarize_rpu()` function in `entrypoint.sh` was implemented,
it looped over all `.rpu.bin` files in the given directory. The function
has since been rewritten to summarize a single RPU given as the first
argument to the function (`$1`). This commit will update the README to
reflect the current functionality of `summarize_rpu()`.