Update `extract-rpu` usage for `dovi_tool` 2.3.0
In `dovi_tool` 2.3.0, support for Matroska (mkv) file input to the
`extract-rpu` command is now considered stable. This commit will update
the `entrypoint.sh` script to extract the initial Dolby Vision Profile 7
RPU from the intact MKV file first before checking the Enhancement Layer
type. If the script is to proceed with conversion, either because a MEL
(Minimal Enhancement Layer) is detected, or because `STOP_IF_FEL=0`, the
script will then extract the HEVC track in order to proceed. The benefit
of this change is primarily in reducing the time to FEL detection.

https://github.com/quietvoid/dovi_tool/releases