Add non-root user
Docker containers run as the `root` user by default. However, it is
considered a best practice to run as a non-root user when possible.

This commit will add a non-root user `apps` (UID `568`) and group `apps`
(GID `568`) for this purpose. This commit will also add the `shadow` apk
package to use the Docker-recommended `groupadd` and `useradd` commands.

For the `docker run`, add the `-u`, `--user` option (`-u apps`).

For Docker Compose, add the `user` key to the appropriate service
(`user: apps`).

- https://docs.docker.com/build/building/best-practices/#user
- https://docs.docker.com/reference/cli/docker/container/run/
- https://docs.docker.com/reference/compose-file/services/#user
- https://pkgs.alpinelinux.org/package/edge/community/x86_64/shadow