Check `entrypoint.sh` as BusyBox ash
The `entrypoint.sh` script starts with a `#!/usr/bin/env sh` directive.
The Docker container image is built on Alpine Linux. On Alpine Linux,
BusyBox ash is the default shell and `/bin/sh` is linked to BusyBox ash.
Therefore, the script actually runs with BusyBox ash.

This commit will add a ShellCheck directive to `entrypoint.sh` so it is
checked as BusyBox ash. Shells supported by ShellCheck are not fully
documented (see SC1103) but can be found in the source code.

https://wiki.alpinelinux.org/wiki/Shell_management#Ash_shell
https://github.com/koalaman/shellcheck/wiki/SC1103
https://github.com/koalaman/shellcheck/blob/acc19ea/src/ShellCheck/Data.hs#L157-L172