Add `dovi_tool`+`hdr10plus_tool` version checks
The Dockerfile includes platform-specific installs of `dovi_tool` and
`hdr10plus_tool` from GitHub Releases. It is possible for installation
errors to occur.

This commit will add `dovi_tool --version`+`hdr10plus_tool --version` to
the Dockerfile to verify that they have been installed successfully.
There hasn't been a problem with this so far, and `entrypoint.sh` will
also check for these tools, but it's a good idea to verify anyway.