Use SHA as `org.opencontainers.image.version`
This commit will update the ordering of the list of tags passed to
docker/metadata-action. The docker/metadata-action README explains:

> The first tag in the list (higher priority) will be used as the image
> version for generated OCI label and version output.

The `type=sha` input will be placed first so the label will show the Git
short SHA (commit ID) in the format "sha-1234567". This is more useful
info than the branch name.

https://github.com/docker/metadata-action?tab=readme-ov-file#priority-attribute