Set Docker labels with docker/metadata-action
Labels are used to set up `--label` options for Buildx and
docker/build-push-action. `--label` options will override `LABEL`
directives in the Dockerfile, and there doesn't seem to be a way to
select only certain labels to override with docker/metadata-action.
Therefore, it's more useful to set labels with docker/metadata-action
and docker/build-push-action than with the Dockerfile.

https://docs.docker.com/build/ci/github-actions/annotations/
https://docs.docker.com/reference/cli/docker/buildx/build/
https://github.com/docker/metadata-action/issues/243
https://github.com/docker/metadata-action?tab=readme-ov-file#global-expressions
https://github.com/docker/build-push-action/blob/9e436ba/src/context.ts#L154-L156